backend/requirements.txt
Django==3.2.3
django-import-export==3.3.1
djangorestframework==3.12.4
djoser==2.1.0
django-colorfield==0.10.1
django-filter==2.3.0
drf-base64==2.0
Pillow==9.0.0
django-link-shortener==0.5